+++ p2panda-net/src/connection/mod.rs
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod actor;
-mod handler;
-mod manager;
-
-pub use actor::{ConnectionActor, ToConnectionActor};
-pub use handler::{SyncConnection, SYNC_CONNECTION_ALPN};